+++ tests/Core/Validator/EventListener/ValidateListenerTest.php
@@ -16,10 +16,10 @@ namespace ApiPlatform\Core\Tests\Validator\EventListener;
 use ApiPlatform\Core\Metadata\Resource\Factory\ResourceMetadataFactoryInterface;
 use ApiPlatform\Core\Metadata\Resource\ResourceMetadata;
 use ApiPlatform\Core\Tests\ProphecyTrait;
-use ApiPlatform\Symfony\EventListener\Exception\ValidationException;
 use ApiPlatform\Symfony\EventListener\ValidateListener;
-use ApiPlatform\Symfony\EventListener\ValidatorInterface;
 use ApiPlatform\Tests\Fixtures\DummyEntity;
+use ApiPlatform\Validator\Exception\ValidationException;
+use ApiPlatform\Validator\ValidatorInterface;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Symfony\Component\HttpFoundation\Request;
